An Online Verification Based on a Real-time Photo by Third Party
PC server and client functions are realized by Java/MySQL, and Android
as well. The third party authentication, which needs a user to take and
send a real-time face image to online friends list asked for a
authentication rather than memorizing and typing alphanumeric password
strings, guarantees a safer and more convenient way of verification.